+++ test/Armada.Test.Unit/Program.cs
@@ -257,6 +257,7 @@ namespace Armada.Test.Unit
             runner.AddSuite(new CheckRunIsolatedCheckoutTests());
             runner.AddSuite(new SiblingLeaseRegistryTests());
             runner.AddSuite(new ConsumerVesselResolverTests());
+            runner.AddSuite(new VoyageCheckArmingPlanTests());
             runner.AddSuite(new DiskLifecycleTests());
             runner.AddSuite(new CheckRunParsingServiceTests());
             runner.AddSuite(new MergeRefCandidateTests());